property cannot get null, even in strict-types
