+++ src/Analyser/NodeScopeResolver.php
@@ -5628,10 +5628,13 @@ final class NodeScopeResolver
 					$assignedNativeType = $scope->getNativeType($assignedExpr);
 					$propertyNativeType = $propertyReflection->getNativeType();
 
+					if ($propertyReflection->hasNativeType() && $propertyNativeType->isNull()->no()) {
+						$assignedExprType = TypeCombinator::removeNull($assignedExprType);
+						$assignedNativeType = TypeCombinator::removeNull($assignedNativeType);
+					}
+
 					if ($propertyReflection->hasNativeType() && $scope->isDeclareStrictTypes()) {
 						$scope = $scope->assignExpression($var, TypeCombinator::intersect($assignedExprType->toCoercedArgumentType(true), $propertyNativeType), TypeCombinator::intersect($assignedNativeType->toCoercedArgumentType(true), $propertyNativeType));
-					} elseif ($propertyReflection->hasNativeType() && $propertyNativeType->isNull()->no()) {
-						$scope = $scope->assignExpression($var, TypeCombinator::removeNull($assignedExprType), TypeCombinator::removeNull($assignedNativeType));
 					} else {
 						$scope = $scope->assignExpression($var, $assignedExprType, $assignedNativeType);
 					}
@@ -5702,10 +5705,13 @@ final class NodeScopeResolver
 					$assignedNativeType = $scope->getNativeType($assignedExpr);
 					$propertyNativeType = $propertyReflection->getNativeType();
 
+					if ($propertyReflection->hasNativeType() && $propertyNativeType->isNull()->no()) {
+						$assignedExprType = TypeCombinator::removeNull($assignedExprType);
+						$assignedNativeType = TypeCombinator::removeNull($assignedNativeType);
+					}
+
 					if ($propertyReflection->hasNativeType() && $scope->isDeclareStrictTypes()) {
 						$scope = $scope->assignExpression($var, TypeCombinator::intersect($assignedExprType->toCoercedArgumentType(true), $propertyNativeType), TypeCombinator::intersect($assignedNativeType->toCoercedArgumentType(true), $propertyNativeType));
-					} elseif ($propertyReflection->hasNativeType() && $propertyNativeType->isNull()->no()) {
-						$scope = $scope->assignExpression($var, TypeCombinator::removeNull($assignedExprType), TypeCombinator::removeNull($assignedNativeType));
 					} else {
 						$scope = $scope->assignExpression($var, $assignedExprType, $assignedNativeType);
 					}
